refactor: discord custom ids
