+++ src/app.module.ts
@@ -4,14 +4,13 @@ import { ScheduleModule } from '@nestjs/schedule';
 import { controllers } from 'src/controllers';
 import { BotCommands } from 'src/discord/commands';
 import { BotEvents } from 'src/discord/events';
-import { BotHelpTicket } from 'src/discord/help-ticket';
-import { BotComposeModal, BotEnvModal, BotLogsModal } from 'src/discord/modals';
+import { BotHelpDesk } from 'src/discord/help-desk';
 import { providers } from 'src/repositories';
 import { services } from 'src/services';
 import { DatabaseService } from 'src/services/database.service';
 
 const middleware = [{ provide: APP_PIPE, useValue: new ValidationPipe({ transform: true, whitelist: true }) }];
-const discord = [BotCommands, BotEvents, BotHelpTicket, BotComposeModal, BotEnvModal, BotLogsModal];
+const discord = [BotCommands, BotEvents, BotHelpDesk];
 
 @Module({
   imports: [ScheduleModule.forRoot()],
